move plone 4 related richtext widget hooks to plone4.py
