+++ src/yafowil/plone/common.py
@@ -4,42 +4,6 @@ from yafowil.common import generic_positional_rendering_helper
 from yafowil.utils import UNSET
 from yafowil.utils import cssid
 from yafowil.utils import managedprops
-from zope.component import getUtility
-from zope.component.hooks import getSite
-import pkg_resources
-
-
-###############################################################################
-# Plone TinyMCE hook (Plone 4 only)
-###############################################################################
-
-HAS_TINYMCE = True
-try:
-    pkg_resources.get_distribution('Products.TinyMCE')
-    from Products.TinyMCE.interfaces.utility import ITinyMCE
-except pkg_resources.DistributionNotFound:
-    HAS_TINYMCE = False
-
-
-if HAS_TINYMCE:
-    def tinymce_config(widget, data):
-        request = data.request.zrequest
-        path = request.physicalPathFromURL(request.getURL())
-        context = getSite()
-        # query object by path, if not found, context is site
-        for i in range(1, len(path)):
-            brains = context.portal_catalog(path={
-                'query': '/'.join(path[:-(i)]),
-                'depth': 0})
-            if not brains:
-                continue
-            context = brains[0].getObject()
-            break
-        utility = getUtility(ITinyMCE)
-        config = utility.getConfiguration(context=context, request=request)
-        return config.replace('"', '&#34;')
-
-    factory.defaults['richtext.title'] = tinymce_config
 
 
 ###############################################################################
